requirements.txt
python-dotenv==1.1.0

fastapi==0.115.11
pydantic==2.10.6
uvicorn==0.34.0

it-core-news-lg@https://github.com/explosion/spacy-models/releases/download/it_core_news_lg-3.8.0/it_core_news_lg-3.8.0-py3-none-any.whl
italian-ats-evaluator==3.0.1

langchain==0.3.20
langgraph==0.3.5
langchain-openai==0.3.7